refactor(dashboard): mueve modules/Dashboard a features/dashboard
