+++ src/actions/board/index.ts
@@ -1,8 +1,5 @@
 export * from './getBoardById'
 export * from './updateBoardName'
-export * from './getBoards'
-export * from './createBoard'
-export * from './deleteBoard'
 export * from './getReminders'
 export * from './saveReminders'
 export * from './getUsageHistory'
