+++ src/features/dashboard/index.ts
@@ -1,0 +1,3 @@
+export type { Board } from './model/board'
+export { default as Dashboard } from './ui/Dashboard'
+export { useDashboardQuery } from './hooks/useDashboardQuery'
