+++ src/features/dashboard/ui/Dashboard.tsx
@@ -1,6 +1,6 @@
-import BoardCard from './components/BoardCard'
-import { useDashboardQuery } from './hooks/useDashboardQuery'
-import CreateBoardDialog from './components/CreateBoardDialog'
+import BoardCard from './BoardCard'
+import { useDashboardQuery } from '../hooks/useDashboardQuery'
+import CreateBoardDialog from './CreateBoardDialog'
 import { Spinner } from '@/shared/ui/atoms/spinner'
 import { Layout } from 'lucide-react'
 import { useTheme } from '@/shared/hooks/useTheme'
